fixed JSR and mode 0 results pdp11_hardware.py - plumbed through the address mode, duuh pdp11_ss_ops.py- plumbed through the address mode, duuh pdp11_ssdd_ops.py- plumbed through the address mode, duuh pdp11_other_ops.py - had wrong jump address in jsr test_M9301_YA.py - removed memory dump test_rss_ops.py - moved two tests out of other test file test_RSS.py - deleted
