+++ .github/workflows/python-app.yml
@@ -35,7 +35,6 @@ jobs:
         pytest test_hardware_psw.py
         pytest test_address_modes.py
         pytest test_condition_code_ops.py
-        pytest test_RSS.py
         pytest test_rss_ops.py
         pytest test_ss_ops.py
         pytest test_ssdd_ops.py
