+++ test_M9301_YA.py
@@ -6,11 +6,7 @@ class TestClass():
         """The M9301-YA is designed specifically for PDP-ll/04 and PDP-ll/34 OEM systems."""
         print('test_M9301-YA')
         pdp11 = PDP11()
-        pdp11.ram.dump(0o500, 0o520)
-        pdp11.ram.dump(0o165320, 0o165340)
         run = pdp11Run(pdp11)
         run.run_in_terminal()
-        pdp11.ram.dump(0o500, 0o520)
-        pdp11.ram.dump(0o165320, 0o165340)
 
 # http://ftpmirror.your.org/pub/misc/bitsavers/pdf/dec/pdp11/xxdp/diag_listings/MAINDEC-11-DQM9A-A-D_M9301_ROM_Bootstrap_Jan77.pdf
